+++ src/controller/users/UsersControllerDelete.java
@@ -1,5 +1,6 @@
 package controller.users;
 
+import controller.access.AccessControllerView;
 import model.User;
 
 import javax.jdo.JDOObjectNotFoundException;
@@ -14,20 +15,37 @@ import java.io.IOException;
 public class UsersControllerDelete extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        PersistenceManager pm = controller.PMF.get().getPersistenceManager();
-
-        String userID = request.getParameter("userID");
-
         try{
-            pm.deletePersistent(pm.getObjectById(User.class, userID));
-            request.getSession().setAttribute("serverResponse","{\"color\": \"#26a69a\",\"response\":\"Service deleted sucessfully.\"}");
-        } catch (JDOObjectNotFoundException e){
-            System.err.println("Exception catched -> " + e.getMessage());
+
+            if (AccessControllerView.checkPermission(request.getSession().getAttribute("userID").toString(),request.getRequestURI())){
+
+                PersistenceManager pm = controller.PMF.get().getPersistenceManager();
+
+                String userID = request.getParameter("userID");
+
+                try{
+                    pm.deletePersistent(pm.getObjectById(User.class, userID));
+                    request.getSession().setAttribute("serverResponse","{\"color\": \"#26a69a\",\"response\":\"User deleted sucessfully.\"}");
+                } catch (JDOObjectNotFoundException e){
+                    System.err.println("Exception catched -> " + e.getMessage());
+                }
+
+                pm.close();
+
+                response.sendRedirect("/users");
+
+            } else {
+
+                request.getSession().setAttribute("serverResponse","{\"color\": \"red\",\"response\":\"You don\\'t have permission to delete a user.\"}");
+                response.sendRedirect("/users");
+
+            }
+
+        } catch (NullPointerException e){
+            e.printStackTrace();
+            response.sendRedirect("/");
         }
 
-        pm.close();
-
-        response.sendRedirect("/users");
 
     }
